+++ src/components/List/List.styles.ts
@@ -1,0 +1,15 @@
+import { css } from '@emotion/react';
+
+export const ListContainerStyle = css`
+  display: flex;
+  flex-direction: column;
+  margin: 3.6rem 0 5.2rem 0;
+`;
+
+export const ListStyle = css`
+  display: flex;
+  flex-direction: column;
+  max-height: 39rem;
+  overflow-y: auto;
+  gap: 2rem;
+`;
