+++ src/components/List/OrderList/OrderList.tsx
@@ -1,0 +1,12 @@
+import ListContainer from '../ListContainer';
+import List from '../List';
+
+function OrderList({ children }: { children: React.ReactNode }) {
+  return (
+    <ListContainer>
+      <List>{children}</List>
+    </ListContainer>
+  );
+}
+
+export default OrderList;
